Ajout du système de quête (prémisse)
-1 ère quête sur la plage disponible (à tester encore plus amplement)
-Création de l'affichage des quêtes (à adapter encore à la résolution)
-Plus généralement création des scripts nécessaires aux quêtes
PS : Roger est pour l'instant une tête volante)

+++ Assets/Scripts/MobSpawners.cs
@@ -8,12 +8,12 @@ using Random = UnityEngine.Random;
 public class MobSpawners : MonoBehaviour
 {
     public GameObject mob;
-    public int x;
-    public int y;
-    public int max_x; //Il faut renseigner dans unity les coordonnées des 4 coins du carré représetant la zone de spawn;
-    public int min_x;
-    public int max_y;
-    public int min_y;
+    public float x;
+    public float y;
+    public float max_x; //Il faut renseigner dans unity les coordonnées des 4 coins du carré représetant la zone de spawn;
+    public float min_x;
+    public float max_y;
+    public float min_y;
     public int enemyCounter;
     public int enemyMaxCount; //Le nombre d'enemy qu'on veut au maximum dans une zone qui peut être variable selon la zone;
     private int _time = 1; //sinon y fait plusieurs boucles en pa
